# CarsMatter/openapi.yaml
swagger: '2.0'
info:
  version: v1
  title: Cars Matter API
host: carfinderai.appspot.com
paths:
  /api/cars/brands:
    get:
      tags:
        - Cars
      operationId: GetAllBrands
      consumes: []
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: Success
          schema:
            uniqueItems: false
            type: array
            items:
              $ref: '#/definitions/Brand'
  /api/cars/brands/models:
    get:
      tags:
        - Cars
      operationId: GetModelsForBrand
      consumes: []
      produces:
        - application/json
      parameters:
        - name: brandHttpPath
          in: query
          required: false
          type: string
      responses:
        '200':
          description: Success
          schema:
            uniqueItems: false
            type: array
            items:
              $ref: '#/definitions/BrandModel'
  /api/cars/brands/models/cars:
    get:
      tags:
        - Cars
      operationId: GetCarsForModel
      consumes: []
      produces:
        - application/json
      parameters:
        - name: modelHttpPath
          in: query
          required: false
          type: string
      responses:
        '200':
          description: Success
          schema:
            uniqueItems: false
            type: array
            items:
              $ref: '#/definitions/Model'
  /api/cars/brands/models/cars/modifications:
    get:
      tags:
        - Cars
      operationId: GetAllCarsModifications
      consumes: []
      produces:
        - application/json
      parameters:
        - name: carModificationsHttpPath
          in: query
          required: false
          type: string
      responses:
        '200':
          description: Success
          schema:
            uniqueItems: false
            type: array
            items:
              $ref: '#/definitions/Model'
  /api/cars/brands/models/cars/image:
    get:
      tags:
        - Cars
      operationId: GetImageForModel
      consumes: []
      produces:
        - application/json
      parameters:
        - name: carImagePath
          in: query
          required: false
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: string
  /api/consumables_notes:
    get:
      tags:
        - ConsumablesNotes
      operationId: GetAllConsumablesNotes
      consumes: []
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: Success
          schema:
            uniqueItems: false
            type: array
            items:
              $ref: '#/definitions/RefillNote'
    put:
      tags:
        - ConsumablesNotes
      operationId: UpdateConsumablesNote
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces:
        - application/json
      parameters:
        - name: consumablesNote
          in: body
          required: false
          schema:
            $ref: '#/definitions/ConsumablesNote'
      responses:
        '200':
          description: Success
          schema:
            type: boolean
    post:
      tags:
        - ConsumablesNotes
      operationId: AddConsumablesNote
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces:
        - application/json
      parameters:
        - name: consumablesNote
          in: body
          required: false
          schema:
            $ref: '#/definitions/ConsumablesNote'
      responses:
        '200':
          description: Success
          schema:
            type: boolean
  '/api/consumables_notes/{id}':
    delete:
      tags:
        - ConsumablesNotes
      operationId: DeleteConsumablesNote
      consumes: []
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          schema:
            type: boolean
  /api/refill_notes:
    get:
      tags:
        - RefillNotes
      operationId: GetAllRefillNotes
      consumes: []
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: Success
          schema:
            uniqueItems: false
            type: array
            items:
              $ref: '#/definitions/RefillNote'
    put:
      tags:
        - RefillNotes
      operationId: UpdateRefillNote
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces:
        - application/json
      parameters:
        - name: refillNote
          in: body
          required: false
          schema:
            $ref: '#/definitions/RefillNote'
      responses:
        '200':
          description: Success
          schema:
            type: boolean
    post:
      tags:
        - RefillNotes
      operationId: AddRefillNote
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces:
        - application/json
      parameters:
        - name: refillNote
          in: body
          required: false
          schema:
            $ref: '#/definitions/RefillNote'
      responses:
        '200':
          description: Success
          schema:
            type: boolean
  '/api/refill_notes/{id}':
    delete:
      tags:
        - RefillNotes
      operationId: DeleteRefillNote
      consumes: []
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          schema:
            type: boolean
definitions:
  Brand:
    type: object
    properties:
      httpPath:
        type: string
      brandName:
        type: string
  BrandModel:
    type: object
    properties:
      httpPath:
        type: string
      modelName:
        type: string
  Model:
    type: object
    properties:
      httpPath:
        type: string
      modelName:
        type: string
      carImagePath:
        type: string
      lowPrice:
        format: double
        type: number
      highPrice:
        format: double
        type: number
      bodyType:
        format: int32
        enum:
          - 0
          - 1
          - 2
          - 3
          - 4
          - 5
          - 6
          - 7
          - 8
        type: integer
  RefillNote:
    type: object
    properties:
      id:
        format: int32
        type: integer
      location:
        type: string
      petrol:
        format: float
        type: number
      odo:
        format: int32
        type: integer
      price:
        format: float
        type: number
      date:
        format: date-time
        type: string
  ConsumablesNote:
    type: object
    properties:
      id:
        format: int32
        type: integer
      date:
        format: date-time
        type: string
      kindOfService:
        type: string
      price:
        format: float
        type: number
      odo:
        format: int32
        type: integer
      location:
        type: string
      notes:
        type: string